+++ src/main/java/org/realityforge/replicant/shared/transport/ReplicantContext.java
@@ -3,7 +3,7 @@ package org.realityforge.replicant.shared.transport;
 public class ReplicantContext
 {
   public static final String SESSION_ID_KEY = "SessionID";
-  public static final String JOB_ID_KEY = "JobID";
+  public static final String REQUEST_ID_KEY = "RequestID";
 
   public static final String REQUEST_ID_HEADER = "X-GWT-RequestID";
   public static final String SESSION_ID_HEADER = "X-GWT-SessionID";